requirements.txt
streamlit
nbformat
google-api-python-client
google-auth
google-auth-oauthlib
google-generativeai
protobuf<4
